% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw_design}
\alias{draw_design}
\title{Design-specific draw method}
\usage{
draw_design(design, data)
}
\arguments{
\item{design}{A design object.}

\item{data}{The data to sample from.}
}
\value{
The sampled rows.
}
\description{
Design-specific draw method
}
\keyword{internal}
