% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design-simple.R
\name{design_simple}
\alias{design_simple}
\title{Simple random sampling}
\usage{
design_simple(n, replace = FALSE)
}
\arguments{
\item{n}{Number of rows to draw. A single non-negative whole number.}

\item{replace}{Sample with replacement?}
}
\value{
A design object, for use with [draw()].
}
\description{
Draws `n` rows uniformly at random.
}
\examples{
df <- data.frame(id = 1:100)
draw(df, design_simple(n = 10), seed = 1)
draw(df, design_simple(n = 150, replace = TRUE), seed = 1)

}
\seealso{
[draw()]

Other designs: 
\code{\link{design_bootstrap}()},
\code{\link{design_cluster}()},
\code{\link{design_multistage}()},
\code{\link{design_reservoir}()},
\code{\link{design_spatial}()},
\code{\link{design_stratified}()},
\code{\link{design_systematic}()},
\code{\link{design_temporal}()},
\code{\link{design_weighted}()}
}
\concept{designs}
