% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{is_design}
\alias{is_design}
\title{Is this a sampling design?}
\usage{
is_design(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if \code{x} was built by one of the \verb{design_*()} constructors.
}
\description{
Is this a sampling design?
}
\examples{
is_design(design_simple(n = 10))
is_design(mtcars)
}
